requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv